# internal/docs/swagger.yaml
basePath: /
definitions:
  internal_modules_order_controller.Order:
    properties:
      complete:
        type: boolean
      id:
        minimum: 1
        type: integer
      petId:
        type: integer
      quantity:
        type: integer
      shipDate:
        type: string
      status:
        type: string
    required:
    - complete
    - id
    - petId
    - quantity
    - shipDate
    - status
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: |-
    Зоомагазин, который работает с тремя основными сущностями: пользователь, животное и магазин.
    Здесь доступны CRUD-операции над всеми сущностями, мягкая работа с пользователем (проставление дат действий), авторизация и выход из системы с помощью чёрного списка, загрузка изображения и сохранение локально на сервере. Это самые интересные возможности, попробуйте сами
  title: Petstore
  version: "1.0"
paths:
...
  /store/order:
    post:
      consumes:
      - application/json
      description: Создание заказа на животное. Отсчёт по id (не должен повторяться).
        Нулевые значения у полей integer не допускаются
      parameters:
      - description: Заполни все поля для размещения заказа
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/internal_modules_order_controller.Order'
      produces:
      - text/plain
      responses:
        "200":
          description: Успешное размещение заказа или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Разместить заказ на животное
      tags:
      - order
  /store/order/{orderId}:
    delete:
      consumes:
      - application/json
      description: Удалить заказ по Id
      parameters:
      - description: Введите id заказа
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Успешное удаление заказа или не ошибочное сообщение
          schema:
            type: string
        "400":
          description: Неверной формат структуры
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Удалить заказ по Id
      tags:
      - order
    get:
      consumes:
      - application/json
      description: Получить заказ по Id
      parameters:
      - description: Введите id заказа
        in: path
        name: orderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Структура заказа или не ошибочное сообщение
          schema:
            $ref: '#/definitions/internal_modules_order_controller.Order'
        "400":
          description: Неверной формат структуры
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Получить заказ по Id
      tags:
      - order
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"